# compute sumw2 in one shot if possible