When we make histograms in one-shot, we can just set sumw2 equal to the bincounts after filling. This gives a 10-20% improvement in speed:

`a,b = randn(10^6),randn(10^6)`
`Hist1D(a, -3:0.01:3)` from 2.8->2.5ms (median)
`Hist2D((a,b),(-3:0.5:3,-3:0.1:3))` from 11.5->9.3ms (median)

If there's a faster way to do `h.sumw2 .= bincounts(h)`, let me know. Then we can potentially be 20-30% faster :)